+++ src/emulator/apphosting/serve.ts
@@ -8,6 +8,7 @@ import { checkListenable } from "../portUtils";
 import { discoverPackageManager } from "./utils";
 import { DEFAULT_HOST, DEFAULT_PORTS } from "../constants";
 import { wrapSpawn } from "../../init/spawn";
+import { getLocalAppHostingConfiguration } from "./config";
 
 /**
  * Spins up a project locally by running the project's dev command.
@@ -32,8 +33,12 @@ export async function start(): Promise<{ hostname: string; port: number }> {
 async function serve(port: number): Promise<void> {
   const rootDir = process.cwd();
   const packageManager = await discoverPackageManager(rootDir);
+  const apphostingLocalConfig = await getLocalAppHostingConfiguration(rootDir);
 
-  await wrapSpawn(packageManager, ["run", "dev"], rootDir, { PORT: port });
+  await wrapSpawn(packageManager, ["run", "dev"], rootDir, {
+    ...apphostingLocalConfig.environmentVariables,
+    PORT: port,
+  });
 }
 
 function availablePort(host: string, port: number): Promise<boolean> {